+++ atom_openmm/__init__.py
@@ -1,0 +1,6 @@
+from importlib.metadata import version, PackageNotFoundError
+
+try:
+    __version__ = version("atom-openmm")
+except PackageNotFoundError:
+    pass
