+++ atom_openmm/utils/singal_guard.py
