+++ atom_openmm/utils/timer.py
